handleIncomingMessages won't use stale socket
It was possible--but highly unlikely without the possiblity of blocking
a socket read--that the handleIncomingMessages thread would read from a
socket which was not the one it was created to read from. If its socket
is closed, and then a LocalEndPoint connects to the same RemoteEndPoint,
the socket handle may be reused, and data belonging to a different
EndPointPair will be stolen. This came up with the blocking behavior,
because the handleIncomingMessages thead would not necessarily read from
its socket soon after it was closed (for instance, by a call to
apiClosedEndPoint).