Example of useful parametric QDisc
The ConnectionBuffers QDisc gives non-standard events: when a connection
is opened, a buffer for the connection is created and passed on. The
EndPoint never gives a Received event. Instead, you'll probably spawn a
thread to clear the connection's input buffer. This is useful because
the rate at which that buffer is cleared controls the rate at which data
is read from the socket to the peer which supplies it. If some handler
is choking up, eventually the peer will have to slow down.