Explicitly use Word32, avoid use of Enum
This patch eliminates the use of signed integers in the protocol and
makes the text more explicit about what is being decoded from the wire.
recvInt32 is not so clear: it was actually used to receive a Word32
almost always, except in the case of control headers and connection
request responses, which would read a signed Int and pass it to
tryToEnum.